pyproject.toml
[project]
name = "weather-assistant"
version = "0.1.0"
description = "Weather Assistant with Claude AI"
dependencies = [
    "streamlit",
    "anthropic",
    "python-dotenv",
    "mcp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
